+++ Assets/Scripts/State.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0d8a4bb13025b4e44afdfd28d2fd51a0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
